move kmt compiler to separate file
