+++ tasks/kmt.py
@@ -14,6 +14,10 @@ from .kernel_matrix_testing.init_kmt import (
 from .kernel_matrix_testing.kmt_os import get_kmt_os
 from .kernel_matrix_testing.tool import Exit, ask, info, warn
 from .system_probe import EMBEDDED_SHARE_DIR
+from .kernel_matrix_testing.compiler import (
+    build_compiler as build_cc,
+    start_compiler as start_cc,
+)
 
 try:
     from tabulate import tabulate
@@ -126,6 +130,13 @@ def revert_resources(ctx):
 
     info("[+] Reverted successfully")
 
+@task
+def build_compiler(ctx):
+    build_cc(ctx)
+
+@task
+def start_compiler(ctx):
+    start_cc(ctx)
 
 def get_vm_ip(stack, version, arch):
     with open(f"{get_kmt_os.stacks_dir}/{stack}/stack.output", 'r') as f:
@@ -214,49 +225,6 @@ def sync(ctx, vms, stack=None, ssh_key=""):
     info("[*] Beginning sync...")
 
     sync_source(ctx, target_vms, "./", "/datadog-agent", ssh_key)
-
-
-def compiler_built(ctx):
-    res = ctx.run("docker images kmt:compile | grep -v REPOSITORY | grep kmt", warn=True)
-    return res.ok
-
-
-@task
-def build_compiler(ctx):
-    ctx.run("docker rm -f $(docker ps -aqf \"name=kmt-compiler\")", warn=True, hide=True)
-    ctx.run("docker image rm kmt:compile", warn=True, hide=True)
-    ctx.run("docker build -f ../datadog-agent-buildimages/system-probe_x64/Dockerfile -t kmt:compile .")
-
-
-def docker_exec(ctx, cmd, user="root"):
-    ctx.run(f"docker exec -u {user} -i kmt-compiler bash -c \"{cmd}\"")
-
-
-@task
-def start_compiler(ctx):
-    if not compiler_built(ctx):
-        build_compiler(ctx)
-
-    if not compiler_running(ctx):
-        ctx.run(
-            "docker run -d --restart always --name kmt-compiler --mount type=bind,source=./,target=/datadog-agent kmt:compile sleep \"infinity\""
-        )
-
-    uid = ctx.run("getent passwd $USER | cut -d ':' -f 3").stdout.rstrip()
-    gid = ctx.run("getent group $USER | cut -d ':' -f 3").stdout.rstrip()
-    docker_exec(ctx, f"getent group {gid} || groupadd -f -g {gid} compiler")
-    docker_exec(ctx, f"getent passwd {uid} || useradd -m -u {uid} -g {gid} compiler")
-    docker_exec(ctx, f"chown {uid}:{gid} /datadog-agent && chown -R {uid}:{gid} /datadog-agent")
-    docker_exec(ctx, "apt install sudo")
-    docker_exec(ctx, "usermod -aG sudo compiler && echo 'compiler ALL=(ALL) NOPASSWD:ALL' >> /etc/sudoers")
-    docker_exec(ctx, f"install -d -m 0777 -o {uid} -g {uid} /go")
-
-
-def compiler_running(ctx):
-    res = ctx.run("docker ps -aqf \"name=kmt-compiler\"")
-    if res.ok:
-        return res.stdout.rstrip() != ""
-    return False
 
 
 TOOLS_PATH = '/datadog-agent/internal/tools'
